LayoutTests/platform/chromium/html5lib/generated/run-tests1-expected.txt

CONSOLE MESSAGE: line 1: Uncaught SyntaxError: Unexpected token <
CONSOLE MESSAGE: line 1: Uncaught SyntaxError: Unexpected token <
../resources/tests1.dat: PASS

Switching to data: url parser (threaded if available) and re-running...

../resources/tests1.dat: PASS
